Started Homework 3
I made the Homework3 script to contain the math I did for Homework 3. I also took the definite integral of the Horton equation on paper to calculate the maximum amount infiltrated -- assuming that the infiltration rate is at its max and is , therefore, the infiltration capacity -- between 2 timepoints, and created this function in the infiltration.py module. I also modified the name of the original integrated function to account for the fact that this function only calculates the total amount infiltrated after a given time rather than between 2 time points.