Plot Green-Ampt infiltration rate vs time
I modified infiltration.py slightly by removing the if-elif-else statements, as making the plot was initially a pain and I kept running into errors with these statements.

To make the plot, I generated a numpy 1-dimensional array of values of F between when ponding started and when the storm ended. I used this array to calculate (1) the times when any of the values in F are reached and (2) the infiltration rate given a specific value of F. I also made 2 more arrays: a 50-value array of the rainfall rate (which is the infiltration rate prior to ponding) and a 50-value array between when the storm started and when ponding started. I then concatenate the arrays that represent values before ponding and the arrays that represent values after ponding together, and then plotted them using matplotlib. I'm quite proud of the result.