Solve for the total amount of infiltration by the end of the storm
The root finding algorithms of scipy.optimize is central to my solving for this value. I created a new function in infiltration.py called stormEnd() which is essentially the same as time() but it subtracts the endingTime, the storm duration. stormEnd() is then fed into another function in my homework script where the only argument that this new function takes is the total amount of infiltration by the end of the storm. This new function is then plugged into scipy.optimize.root to find its root, which is the total amount of infiltration by the end of the storm.