Created a method in infiltration.py for making graph data
The function that I created is called graphData() and it produces a 2-D numpy array of time and infiltration rate values as calculated using the Green-Ampt model. This array can then be graphed in a plot of infiltration rate vs time. I created this function to further abstract the steps to plot infiltration rates vs time according to the Green-Ampt model. I replaced the chunks of code in Homework3.py and Week7_discussion.py scripts that are used to make arrays of values to graph with 1 or 2 lines of code that calls this method.